django_proj/requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
coverage==4.5.3
Django==1.11.18
django-admin-sortable2==0.6.21
django-ajax-selects==1.7.1
django-auditlog==0.4.5
django-cors-headers==2.4.0
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.1.4
django-guardian==1.4.9
django-ipware==2.1.0
django-jsonfield==1.2.0
django-lookup-tables==0.15.3
django-model-utils==3.1.2
django-modeladmin-reorder==0.3.1
django-postgrespool==0.3.1
django-robots==3.1.0
django-sendgrid-v5==0.8.0
django-storages==1.7.1
django-taggit==0.23.0
django-taggit-serializer==0.1.7
django-treebeard==4.3
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.5
djangorestframework==3.9.0
entrypoints==0.3
flake8==3.7.7
future==0.17.1
gevent==1.4.0
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.18
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
mimeparse==0.1.3
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
protobuf==3.8.0
psycopg2==2.8.3
psycopg2-binary==2.7.5
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.0
PySCIPOpt==2.1.3
pytest==4.6.3
pytest-cov==2.7.1
pytest-django==3.5.0
pytest-mock==1.10.4
pytest-pythonpath==0.7.3
python-dateutil==2.6.0
python-http-client==3.1.0
pytz==2018.9
requests==2.21.0
rsa==4.0
sendgrid==6.0.5
six==1.12.0
SQLAlchemy==1.3.5
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.24.3
wcwidth==0.1.7
whitenoise==3.3.1
wrapt==1.11.2
xlrd==1.2.0
zipp==0.5.1
